Adding more to Bakufu, CameraShaking, and CloudManagement!
As for Camera Shake, there is a Camera Shake Manager object that controls the camera shake, and a new script that controls the time intervals in which the screen shakes. This will be used for the tremor and earthquake effect later; the length and intensity of the camera shaking should change depending on whether it is an earthquake or a tremor. As for Cloud Management, I have added in a CloudManager script that will be attached to each cloud that should move. It has a public variable which can change the speed of the cloud. In addition, the cloud will reset to its original position if it enters a CloudEnd-tagged collider with a rigidbody.